+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - webui: optimize performance and security [PR #2628]
 - matrix: add ubuntu 26.04 [PR #2636]
 - python-bareos: add support for the sslpsk3 module [PR #2637]
+- python-dir: fix not registering python thread state correctly#2493 [PR #2605]
 
 ## [25.0.3] - 2026-04-01
 
@@ -2266,6 +2267,7 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2598]: https://github.com/bareos/bareos/pull/2598
 [PR #2602]: https://github.com/bareos/bareos/pull/2602
 [PR #2604]: https://github.com/bareos/bareos/pull/2604
+[PR #2605]: https://github.com/bareos/bareos/pull/2605
 [PR #2606]: https://github.com/bareos/bareos/pull/2606
 [PR #2608]: https://github.com/bareos/bareos/pull/2608
 [PR #2617]: https://github.com/bareos/bareos/pull/2617